+++ numpy/linalg/lapack_lite/wrapped_routines
@@ -46,5 +46,4 @@ zpotrs
 zungqr
 # need this b/c it's not properly declared as external in the BLAS source
 dcabs1
-IGNORE: dlamch
 IGNORE: xerbla
